feat(HRtoCVR): move HRClients to folder + update namespace
